Changed call_price float type to decimal type

+++ call_project/calls/models.py
@@ -98,7 +98,8 @@ class Bill(models.Model):
                                null=True, blank=False, unique=True)
     call_start_date = models.DateField(verbose_name="Data da Ligação")
     call_start_time = models.TimeField(verbose_name="Horário da Ligação")
-    call_price = models.FloatField(verbose_name="Valor da chamada")
+    call_price = models.DecimalField(verbose_name="Valor da chamada",
+                                     decimal_places=2, max_digits=10)
     duration = models.TimeField(verbose_name="Duração da Ligação",
                                 null=True)
